Fix draft auto-saving
The auto-save handlers do not pass an event to the save method, so calling event.preventDefault(); unconditionally will fail.
Solution? Add a conditional, duh.

Now, when was this broken? I'm not sure.
git blame says that commit 5bd201eab51677b6f5ce14b08f053bda0bc5705b (June 2014) added the param and the event.preventDefault(); call but I'm fairly confident that the auto-save was working for me locally circa/after 2016 or 2017 or so.

Change-Id: I78cb6742fbb6c2e9ba7e79a28978a2be1ce157c8