Maybe fix weird fatal related to there not being a valid Title object + two more fatals related to now-private EditPage properties
The previous logic made no sense. First we checked that whether wpDraftTitle was null, and if it was, we'd then proceed to...call Title::newFromText() on that a few line below?!
WTF.
Of course that's gonna fail, but nothing here handles that, and that's still true.
However, now we also check for the regular MW 'title' URL param if 'wpDraftTitle' is not there and hope that one or the other can be used to make a bit more robust Title since Draft#save requires a valid Title.

This can still fail and there should be proper checking for invalid Title object somewhere in here, but this oughta be *slightly* more robust than the previous version.
We now explicitly check for the "non-save submissions", i.e. whether either the "Show preview" or "Show changes" button was pressed.
TODO: Are there more things to check for in there?

Additionally fix calls to two now-private (since MW 1.38) EditPage properties, which were preventing loading a draft from Special:Drafts due to PHP fatals

Change-Id: I4695f760c99387923dd8a0cc3ac4c5f4ef5c0a35